UpdateUserRequest

+++ routes/api.php
@@ -17,17 +17,17 @@ use App\Http\Controllers\MediaController;
 });*/
 
 //Pruebas en postman incluir /api/.....
-Route::post('/user', [UserController::class, 'store']);//ok
+Route::post('/register', [UserController::class, 'store']);//ok
 //Para login: http://localhost/api/user/login (añadir Bearer Token y Content-Type)
 Route::post('/login', [UserController::class, 'verify']);//ok
 
 /************* Dentro del grupo las rutas que estan protegidas con middleware.*******************/
 Route::middleware('auth:sanctum')->group(function () {
 
-    Route::get('/users', [UserController::class, 'index']);
-    Route::get('/user/{user}', [UserController::class, 'show']);
+    Route::get('/users', [UserController::class, 'index']);//ok admin + token.
+    Route::get('/user/{user}', [UserController::class, 'show']);//ok admin, y user + token.
     Route::put('/user/{user}', [UserController::class, 'update']);
-    Route::delete('/users/{user}', [UserController::class, 'destroy']);
+    Route::delete('/users/{user}', [UserController::class, 'destroy']);//ok admin y authUser/token.
     Route::post('/user/logout', [UserController::class, 'logout']);//ok
 
     //Rutas REST para Pieza.
